requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile requirements.in
#
annotated-types==0.7.0
    # via pydantic
anyio==4.9.0
    # via
    #   httpx
    #   starlette
    #   watchfiles
bcrypt==4.0.1
    # via -r requirements.in
certifi==2025.1.31
    # via
    #   httpcore
    #   httpx
    #   requests
charset-normalizer==3.4.1
    # via requests
click==8.1.8
    # via
    #   rich-toolkit
    #   typer
    #   uvicorn
colorama==0.4.6
    # via
    #   click
    #   tqdm
    #   uvicorn
dnspython==2.7.0
    # via email-validator
ecdsa==0.19.1
    # via python-jose
email-validator==2.2.0
    # via fastapi
fastapi[standard]==0.115.11
    # via -r requirements.in
fastapi-cli[standard]==0.0.7
    # via fastapi
greenlet==3.1.1
    # via sqlalchemy
h11==0.14.0
    # via
    #   httpcore
    #   uvicorn
httpcore==1.0.7
    # via httpx
httptools==0.6.4
    # via uvicorn
httpx==0.28.1
    # via fastapi
idna==3.10
    # via
    #   anyio
    #   email-validator
    #   httpx
    #   requests
jinja2==3.1.6
    # via fastapi
markdown-it-py==3.0.0
    # via rich
markupsafe==3.0.2
    # via jinja2
mdurl==0.1.2
    # via markdown-it-py
numpy==2.1.1
    # via
    #   -r requirements.in
    #   opencv-python
    #   pandas
opencv-python==4.11.0.86
    # via -r requirements.in
packaging==24.2
    # via pytesseract
pandas==2.2.3
    # via -r requirements.in
passlib==1.7.4
    # via -r requirements.in
pillow==11.1.0
    # via
    #   -r requirements.in
    #   pytesseract
psycopg2==2.9.10
    # via -r requirements.in
pyasn1==0.4.8
    # via
    #   python-jose
    #   rsa
pydantic==2.10.6
    # via fastapi
pydantic-core==2.27.2
    # via pydantic
pygments==2.19.1
    # via rich
pytesseract==0.3.13
    # via -r requirements.in
python-dateutil==2.9.0.post0
    # via pandas
python-dotenv==1.1.0
    # via uvicorn
python-jose==3.4.0
    # via -r requirements.in
python-multipart==0.0.20
    # via fastapi
pytz==2025.1
    # via pandas
pyyaml==6.0.2
    # via uvicorn
pyzbar==0.1.9
    # via -r requirements.in
requests==2.32.3
    # via -r requirements.in
rich==13.9.4
    # via
    #   rich-toolkit
    #   typer
rich-toolkit==0.14.0
    # via fastapi-cli
rsa==4.9
    # via python-jose
shellingham==1.5.4
    # via typer
six==1.17.0
    # via
    #   ecdsa
    #   python-dateutil
sniffio==1.3.1
    # via anyio
sqlalchemy==2.0.39
    # via -r requirements.in
starlette==0.46.1
    # via fastapi
tqdm==4.67.1
    # via -r requirements.in
typer==0.15.2
    # via fastapi-cli
typing-extensions==4.12.2
    # via
    #   anyio
    #   fastapi
    #   pydantic
    #   pydantic-core
    #   rich-toolkit
    #   sqlalchemy
    #   typer
tzdata==2025.1
    # via pandas
urllib3==2.3.0
    # via requests
uvicorn[standard]==0.34.0
    # via
    #   fastapi
    #   fastapi-cli
watchfiles==1.0.4
    # via uvicorn
websockets==15.0.1
    # via uvicorn
